Dovrseni repozitorijumi i servisi, ispravljeni interfejsi i convertor klase
Repozitorijumi i servisi su sada funkcionalni, interfejsi rade sa
integerima umesto long, convertor klase ispravljene zbog izmena imena
methoda u servisima

+++ src/main/java/com/pi/PoslovnaInformatika/repository/FakturaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.pi.PoslovnaInformatika.model.Faktura;
 
 
-public interface FakturaRepository extends JpaRepository<Faktura, Long> {
+public interface FakturaRepository extends JpaRepository<Faktura, Integer> {
 
 }
