+++ src/main/java/com/pi/PoslovnaInformatika/repository/NarudzbenicaRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.pi.PoslovnaInformatika.model.Narudzbenica;
 
 
-public interface NarudzbenicaRepository extends JpaRepository<Narudzbenica, Long> {
+public interface NarudzbenicaRepository extends JpaRepository<Narudzbenica, Integer> {
 
 }
